requirements.txt
annotated-types==0.7.0
anyio==4.10.0
azure-core==1.35.0
azure-cosmos==4.9.0
azure-functions==1.23.0
azure-identity==1.24.0
azurefunctions-extensions-base==1.0.0b2
azurefunctions-extensions-http-fastapi==1.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.74
langchain-openai==0.3.29
langgraph==0.6.4
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.0
langsmith==0.4.13
MarkupSafe==3.0.2
msal==1.33.0
msal-extensions==1.3.1
openai==1.99.9
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyodbc==5.2.0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==3.1.3
xxhash==3.5.0
zstandard==0.23.0
